Cleanup and README.md refactoring
